@@ -1,12 +1,14 @@
 <?php
-namespace PhitFlyer;
+declare(strict_types=1);
 
-use NetDriver\Http\HttpRequest;
-use NetDriver\NetDriverInterface;
+namespace Stk2k\PhitFlyer;
+
+use Stk2k\NetDriver\Http\HttpRequest;
+use Stk2k\NetDriver\NetDriverInterface;
 
 use Psr\Log\LoggerInterface;
 
-use PhitFlyer\Exception\PhitFlyerClientExceptionInterface;
+use Stk2k\PhitFlyer\Exception\PhitFlyerClientExceptionInterface;
 
 /**
  * Logger decorator
